Make selector private
